feat: implement basic watchlist features
This commit enables the user to add not streamable Songs to a watchlist.
The user specific watchlist will recheck the songs and inform the user
if any of the watched songs changes to beeing streamable.

For this purpose the bot will store the song information from twich, to
reduce API calls.
The discord user ID will also be stored in order to maintain a per user
watchlist.