feat: refactor spotify check
This commit reduces the verbosity for daily usage to a minimum.

Wether a song was found, is beeing processed or (not) streamable can be
found as a reaction.

If a song was not recognized we have to use the alternative songs from
twitch. These will be rendered into a thread, which will be closed after
posting the alternative list.